+++ mobile/android/android-components/components/support/ktx/src/main/java/mozilla/components/support/ktx/android/view/Activity.kt
@@ -7,6 +7,7 @@ package mozilla.components.support.ktx.android.view
 import android.app.Activity
 import android.view.View
 import android.view.WindowManager
+import mozilla.components.support.base.log.logger.Logger
 
 /**
  * Attempts to call immersive mode using the View to hide the status bar and navigation buttons.
@@ -32,4 +33,23 @@ fun Activity.exitImmersiveModeIfNeeded() {
 
     window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
     window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_VISIBLE
+}
+
+/**
+ * Calls [Activity.reportFullyDrawn] while also preventing crashes under some circumstances.
+ *
+ * @param errorLogger the logger to be used if errors are logged.
+ */
+fun Activity.reportFullyDrawnSafe(errorLogger: Logger) {
+    try {
+        reportFullyDrawn()
+    } catch (e: SecurityException) {
+        // This exception is throw on some Samsung devices. We were unable to identify the root
+        // cause but suspect it's related to Samsung security features. See
+        // https://github.com/mozilla-mobile/fenix/issues/12345#issuecomment-655058864 for details.
+        //
+        // We include "Fully drawn" in the log statement so that this error appears when grepping
+        // for fully drawn time.
+        errorLogger.error("Fully drawn - unable to call reportFullyDrawn", e)
+    }
 }