+++ mobile/android/android-components/docs/changelog.md
@@ -4,13 +4,21 @@ title: Changelog
 permalink: /changelog/
 ---
 
-# 51.0.0-SNAPSHOT (In Development)
+# 52.0.0-SNAPSHOT (In Development)
 
-* [Commits](https://github.com/mozilla-mobile/android-components/compare/v50.0.0...master)
-* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/110?closed=1)
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v51.0.0...master)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/111?closed=1)
 * [Dependencies](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Dependencies.kt)
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
+
+# 51.0.0
+
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v50.0.0...v51.0.0)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/110?closed=1)
+* [Dependencies](https://github.com/mozilla-mobile/android-components/blob/v51.0.0/buildSrc/src/main/java/Dependencies.kt)
+* [Gecko](https://github.com/mozilla-mobile/android-components/blob/v51.0.0/buildSrc/src/main/java/Gecko.kt)
+* [Configuration](https://github.com/mozilla-mobile/android-components/blob/v51.0.0/buildSrc/src/main/java/Config.kt)
 
 * **browser-session**
   * ⚠️ **This is a breaking change**: Removed the following properties from `Session`: `findResults`, `hitResult`, `fullScreenMode` and `layoutInDisplayCutoutMode`. The values still exist in `BrowserState` and can be read from there.